Refactor EventLoop singleton to namespace methods
A few things changed in this commit:
- Component renamed from genericeventhandler to event_handlers
The component had generic in its name, but none of the classes beared it, so I removed the "generic" part of the name. Secondly, it was no event loop, it was a list of event handlers, so a simpler name for the class was EventHandlers.

- Refactored singleton class to namespace
The code using this components should not be bothered if the implementation uses a singleton, a factory, or whatever. Using get_instance limits you to refactor what happens internally as you're somewhat exposing your internals. When the instance methods (add, remove, handle for simplicity) are made available at static level you're almost at the same point as using a namespace with methods. Because the list `handlers` is declared in the namespace and in the cpp file, others can't access it or add new handlers suddenly.

- Renamed event_handler to handle_event
The action requested is to handle an event, so I've renamed the method to reflect it being an action.