Message Handler & Dispatcher for Loconet
The Loconet Handler is a singleton inbound message handler. It queues
messages, and allows different handlers to handle upon incoming
messages.
In a similar way, the Dispatcher is a singleton for sending messages
out.

Note that for this implementation, I had to overload the New and Delete
operator of Message, to ensure that they use malloc for memory
allocation. Otherwise, the object would be created within the heap of a
task. This would hamper to share a Message object between tasks, e.g. a
scenario like: 'the handler would take the message in Task A, and then a
handler wants to process it in Task B' would not be possible.

IMPORTANT: this requires extensive testing!