Bugix: crash on accept() of socket
If there is not enough time after sending and the next accept(), the
system crashes on a mutex exception in FreeRTOS (called in arch_sys.h of
lwip). The trick to mitigate for now: have a large enough delay between
the process_sessions() and process_accept(). This delay should be at
least 100 ms.